Port the wasi-filesystem API to the new wai format.
This makes a number of changes, to make use of interface-types features such
as `expected`, variant types, and resources. The change to use resources in
particular means that filesystem functions are now methods of the `descriptor`
resource. Since this means renaming everything, take this opportunity to
introduce a new naming conventions, with `_at` being used for functions that
take dirfd+path arguments.

This also eliminates the `rights` concept what was present in earlier versions
of WASI, has has discussed in #31. This required adding new flags to `open_at`,
so while here, this also adds basic `chmod`-like support, as discussed in #33.

And, this removes support for readdir seeking (seekdir/telldir), as discussed
in #7.

And it adds a fifo file type and a more general socket type, as discussed in