Re-introduce the `badf` filesystem error code.
POSIX overloads `EBADF` with two meanings: file descriptor out of
bounds, and attempt to read or write on a device that doesn't support
reading or writing, respectively. In WASI, the former meaning is
being handled at the WIT level, but the latter meaning still applies,
so reintroduce `badf` as an error code.