Add advisory file locking support.
Add functions which correspond to `flock` in Unix. This is not a POSIX
interface, but it's sufficiently portable, there's a way to implement it
on Windows, and it doesn't have problematic process-associated locks.

Instead of having a single `flock` function that gets passed and enum,
this has separate functions for shared vs. exclusive, and locking vs.
non-locking. That's something we can revisit if needed, but it keeps the
API simpler for now.