Rename `info` to `fd-info`, and other minor fixes. (#66)
* Rename `info` to `fd-info`, and other minor fixes.

Rename the `info` function to `fd-info`, to avoid a name collision with
the `info` type.

While here, also fix the documentation name for the `datasync` function.
And use `filesize` and for fadvise arguments.

* Swap `info` and `fd-info`.

It feels more natural to have the function be named `info` and the
record `fd-info`. `info` is a method of the `descriptor` resource,
which provides a scope, while `fd-info` is just a standalone record.