Split `info` into `type` and `flags`, and add a `set-info`.
Split the `info` function into `type` and `flags` functions, which return
each of the fields.

Remove the `fd-info` type, which is no longer needed, and rename `fd-type`
and `flags` to `descriptor-type` and `descriptor-flags`, respectively.

And add a `set-flags` function, which adds `F_SETFL` functionality which
was lost in the translation to wit.