Change `readdir` from returning bytes to returning a `dir-entry`s.
`readdir` returning bytes has been a source of complexity for both libc
and engine implementors, with subtle issues about alignment and partial
records.

Rename `dirent` to `dir-entry` for readability, and revamp it:
 - Make the name field a proper `string` rather than being represented
   by trailing bytes in the buffer.
 - Make the inode field `optional`, so that we can make it optional
   as discussed in #65 without special-casing zero, which is reportedly
   a valid inode number on [some filesystems].

And remove the `rewind` parameter, which isn't needed when we return
a stream, as users wanting to start at the beginning can just request
a new stream.

[some filesystems]: https://sourceware.org/pipermail/libc-alpha/2022-September/142059.html