Add advisory file locking support. (#69)
* Add advisory file locking support.

Add functions which correspond to `flock` in Unix. This is not a POSIX
interface, but it's sufficiently portable, there's a way to implement it
on Windows, and it doesn't have problematic process-associated locks.

Instead of having a single `flock` function that gets passed and enum,
this has separate functions for shared vs. exclusive, and locking vs.
non-locking. That's something we can revisit if needed, but it keeps the
API simpler for now.

* Document what happens if a lock is already held.

* Add notes about interactions between WASI locks and non-WASI locks.

Also mention that exclusive locks may require write mode on file
descriptors, as described in the [Linux `flock` man page].

[Linux `flock` man page]: https://man7.org/linux/man-pages/man2/flock.2.html

* Add a comment that not all filesystems support locking.

* Update wasi-filesystem.abi.md.