Document the behavior of writes past the end of a file.
Document that seek+write and pwrite can write past the end of the file,
and that the file is extended with zeros as needed.

This will require extra implementation work on Windows, where writes past
the end of files extend with uninitialized bytes, however the upside of
the tradeoff is cleaner and more POSIX-compliant behavior for
applications, which seems worthwhile here.

Also, document that `seek` and `tell` fail on a directory.

Fixes #13.