Add support for readonly directories.
Add a `mutate-directory` flag to `descriptor-flags`, and make it a
requirement for any filesystem-mutating operation. POSIX compatibility
may be provided by setting `mutate-directory` whenever performing a
readonly `open-at`.

When this flag is unset on a directory descriptor, the descriptor acts
as a readonly directory view.

This is the most notable use case that the "rights" system in preview1
was useful for.