Add an `access-at` function.
Adds an `access-at` function which checks for a filesystem path
being readable, writeable, or executable, or whether it exists,
without performing an actual access.

This is similar to `faccessat` in POSIX, though it only supports
the `AT_EACCESS` mode, where the effective user id is tested, as
WASI doesn't otherwise expose the difference between the effective
and real user IDs.