types.wit: delete file locking methods
We believe there is value in these lock methods for some applications
(e.g. sqlite), but for the sake of shipping Preview 2, we don't have
time to make and test a a cross platform implementation of these in two
separate engines at this time. This is in part because some additional
design work is required to integrate the blocking lock functions with
pollable.

These functions may come back as part of an additional filesystem
interface in a post Preview 2 patch version 0.2.n, or in Preview 3 or
beyond. Until then, their design is still in the git history.